requirements.txt
# Backend dependencies
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
numpy==1.25.2


# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Instagram scraping
instaloader==4.10.3

# Ambiente e configuração
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# IA e OpenAI
openai==1.3.7

# Geração de PDFs
reportlab==4.0.7
Pillow==10.1.0

# Utilitários
requests==2.31.0
aiohttp==3.9.1

# Data e tempo
python-dateutil==2.8.2